requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.6
einops==0.8.0
eng-syl==4.0.1
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
gpt4all==2.6.0
grpcio==1.63.0
h5py==3.11.0
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
inquirerpy==0.3.4
iso639==0.1.4
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
keras==3.3.3
langchain==0.2.0
langchain-core==0.2.0
langchain-text-splitters==0.2.0
langcodes==3.4.0
langsmith==0.1.59
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
packaging==23.2
pfzy==0.3.4
pillow==10.3.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
SQLAlchemy==2.0.30
sympy==1.12
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.0
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uroman-python==1.2.8.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.2
